rp-poetry/pyproject.toml
[tool.poetry]
name = "rp-poetry"
version = "0.1.0"
description = ""
authors = ["wellersonpch <wellersonpch1@gmail.com>"]
readme = "README.md"
packages = [{include = "rp_poetry"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
